+++ Backend/src/main/java/com/pastelarte/pastelarte_api/dto/PedidoResponseDTO.java
@@ -1,7 +1,9 @@
+// PedidoResponseDTO.java
 package com.pastelarte.pastelarte_api.dto;
 
 import lombok.*;
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -14,7 +16,7 @@ public class PedidoResponseDTO {
     private String nombreCliente;
     private LocalDate fecha;
     private String estado;
-    private Double total;
+    private BigDecimal total;
     private Integer idPago;
     private String comprobante;
     private String numeroNequi;
